+++ ReleaseNotes.md
@@ -15,6 +15,7 @@
 * TGR-638: Workflow UI: It is now possible the show HTML text in the workflow message 
 * TGR-663: WebUi: All headers are now collapsable
 * TGR-577: User can ignore scenarios or whole features by adding "@Ignore" above the scenario/feature in the feature file
+* TGR-678: The log files in the target/rbellogs directory are now saved with a timestamp at the end of the file name
 
 ## Breaking changes