+++ ReleaseNotes.md
@@ -18,6 +18,8 @@
 * TGR-678: The log files in the target/rbellogs directory are now saved with a timestamp at the end of the file name
 * TGR-596: WebUi: Text based Regex/Search as search filter
 * TGR-683: WebUi: Button for binary content fixed
+* TGR-544: TigerProxy: Added option for TLS-version
+* TGR-463: re-/create TigerProxy ID from client certificate 
 
 ## Breaking changes
 
@@ -26,6 +28,10 @@
 ## Bugfixes
 
 * TGR-662: trafficEndpointFilterString are now honored for standalone tiger proxies
+* TGR-682: reverse-proxy does not delete host-header-field
+* TGR-679: mouse-over for long requests fixed
+* TGR-660: Admin UI fixed
+
 
 # Release 0.24.0
 
@@ -37,7 +43,6 @@
 
 ## Breaking changes
 
-* TGR-540: Migration unto the main-branch mockserver. This breaks the client-address. Rbel-Messages no longer carry the information who sent the request (or who received the response). This will be added back in a later version (Ticket TGR-651)
 * TGR-613: We removed the deprecated Cucumber-Steps in RBelValidatorGlue for good
 
 ## Bugfixes